requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-jose
passlib[bcrypt]
python-multipart
redis
celery
pydantic
email-validator
slowapi
httpx
pytest
pytest-asyncio